pyproject.toml
[project]
name = "digital-employee-demo"
version = "0.1.0"
description = "Digital Employee Memory System Demo - LangGraph + LangMem ReAct Agent"
requires-python = ">=3.11,<3.13"
readme = "README.md"

dependencies = [
    # Agent framework
    "langgraph>=0.2.0",
    "langgraph-checkpoint-postgres>=2.0.0",
    "langmem>=0.0.20",
    "langchain-openai>=0.2.0",
    # LLM & Embedding
    "openai>=1.0.0",
    "httpx>=0.27.0",
    # Search
    "tavily-python>=0.5.0",
    # Web UI
    "chainlit>=1.3.0",
    "streamlit>=1.30.0",
    # API service
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    # Database
    "asyncpg>=0.30.0",
    "psycopg[binary]>=3.2.0",
    "sqlalchemy>=2.0.0",
    "greenlet>=3.0.0",
    # Utilities
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "pillow>=12.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
